+++ creditcardapp/src/main/java/com/banking/creditcardapp/repository/CreditCardDetailsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.banking.creditcardapp.entity.CreditCardDetails;
 
-public interface CreditCardDetailsRepository extends JpaRepository<CreditCardDetails, Integer> {
+public interface CreditCardDetailsRepository extends JpaRepository<CreditCardDetails, Long> {
     Optional<CreditCardDetails> findByCustomerEmail(String email);
 }
